+++ api/system/service/control.py
@@ -204,11 +204,8 @@ class SystemReloadThread(Thread):
         return 'erigonesd:mgmt'
 
     def reload_service(self, name):
-        # Reload gunicorn-gui and restart gunicorn-{sio|api} and erigonesd - issue esdc-ce#20
-        if name == 'app-gui':
-            self.sctrl.reload(name)
-        else:
-            self.sctrl.restart(name)
+        # Issue esdc-ce#20
+        self.sctrl.restart(name)
 
     def reload_all(self):
         last = None