+++ doc/changelog.rst
@@ -1,9 +1,18 @@
 Changelog
 #########
 
-
-2.4.0 (unreleased)
+2.4.1 (unreleased)
 ==================
+
+Features
+--------
+
+Bugs
+----
+
+
+2.4.0 (released on 2017-02-22)
+==============================
 
 Features
 --------
@@ -25,6 +34,8 @@ Features
 - Updated design of DC switch button - `#64 <https://github.com/erigones/esdc-ce/issues/64>`__
 - Changed image repository view to show last 30 images by default - `#66 <https://github.com/erigones/esdc-ce/issues/66>`__
 - Improved consistency and UX of modal button links - `#39 <https://github.com/erigones/esdc-ce/issues/39>`__
+- Modified update script bin/esdc-git-update to fail when git fetch fails and display usage for invalid invocation - `#68 <https://github.com/erigones/esdc-ce/issues/68>`__
+- Removed Linux Zone images from Import images view as it's not supported for now - `#73 <https://github.com/erigones/esdc-ce/issues/73>`__
 
 Bugs
 ----
@@ -34,6 +45,12 @@ Bugs
 - Updated message box that displays information about unavailable nodes to show/hide dynamically - `#35 <https://github.com/erigones/esdc-ce/issues/35>`__
 - Fixed image import of images with same name - `#61 <https://github.com/erigones/esdc-ce/issues/61>`__
 - Fixed initial VM harvest problem with temporary unreachable worker - `#61 <https://github.com/erigones/esdc-ce/issues/61>`__
+- Changed reload to restart of application GUI service - commit `#05f9702 <https://github.com/erigones/esdc-ce/commit/05f97027ac542c4f284892fd3aa85e1576a553ed>`__
+- Fixed redirect after VM hostname change - `#70 <https://github.com/erigones/esdc-ce/issues/70>`__
+- Fixed minor issues in Import/Export functionality - `#71 <https://github.com/erigones/esdc-ce/issues/71>`__
+- Fixed language switching in user profile - `#72 <https://github.com/erigones/esdc-ce/issues/72>`__
+- Fixed ``GET /task/log -page <number>`` API view - `#74 <https://github.com/erigones/esdc-ce/pull/74>`__
+- Fixed object_type filter in Task Log (API & GUI) - `#74 <https://github.com/erigones/esdc-ce/pull/74>`__
 
 
 2.3.3 (released on 2017-02-04)